requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
click==8.1.7
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.0.1
ffmpeg-python==0.2.0
fonttools==4.49.0
future==1.0.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pyzmq==25.1.2
scipy==1.15.2
setuptools==69.1.1
six==1.16.0
spark-parser==1.8.9
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4
traitlets==5.14.2
uncompyle6==3.7.4
wcwidth==0.2.13
wheel==0.42.0
xdis==5.0.13
